Cleaned up file structure in pages, condensed and relocated css file for all pages into styles folder
