+++ src/pages/CollabInfo.js
@@ -1,8 +1,8 @@
 // import { Link } from "react-router-dom";
-import Container from "../../components/Container";
-import Col from "../../components/Col";
-import Row from "../../components/Row";
-import './index.css';
+import Container from "../components/Container";
+import Col from "../components/Col";
+import Row from "../components/Row";
+import '../styles/index.css';
 
 function CollabInfo() {
     return (
