+++ src/pages/Forum.js
@@ -1,8 +1,8 @@
 // import React, { useRef } from "react";
 // import { Link } from "react-router-dom";
-import Container from "../../components/Container";
-import Col from "../../components/Col";
-import './index.css';
+import Container from "../components/Container";
+import Col from "../components/Col";
+import '../styles/index.css';
 
 function Forum() {
     // const titleRef = useRef();
