+++ src/pages/Landing.js
@@ -1,10 +1,10 @@
 // import React, { useState } from "react";
 // import { Link } from "react-router-dom";
-import Container from "../../components/Container";
-import Col from "../../components/Col";
-import Row from "../../components/Row";
-import './index.css';
-import Collab from "../../styles/Collab_Dia_Final.png";
+import Container from "../components/Container";
+import Col from "../components/Col";
+import Row from "../components/Row";
+import '../styles/index.css';
+import Collab from "../styles/Collab_Dia_Final.png";
 
 function Landing() {
