+++ src/pages/ProfileEditor.js
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 // import { Link } from "react-router-dom";
-import Container from "../../components/Container";
-import Col from "../../components/Col";
-import Row from "../../components/Row";
-import './index.css';
+import Container from "../components/Container";
+import Col from "../components/Col";
+import Row from "../components/Row";
+import '../styles/index.css';
 
 function ProfileEditor() {
     const [username, setUsername] = useState();
